FSM allocation
